requirements.txt
chromadb>=0.5.0
sentence-transformers>=2.5.1
pillow>=10.0.0
langchain-community>=0.0.30
langchain-text-splitters>=0.0.1
langchain>=0.1.0
langgraph>=0.0.20
langchain-core>=0.1.0
langchain-openai>=0.0.8
pypdf>=3.10.0